backend/requirements.txt
# Django and REST
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.2
drf-yasg==1.21.5
djangorestframework-simplejwt==5.2.2

# Database and async tasks
psycopg2-binary==2.9.10

celery==5.3.1
redis==4.6.0
python-dotenv==1.0.0
requests==2.31.0

# AI / ML / NLP
transformers==4.31.0
torch==2.0.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.26.4

# Deployment
gunicorn==21.2.0
